config/__init__.py
from dotenv import load_dotenv, find_dotenv
import os
from .settings import *

load_dotenv(find_dotenv(), verbose=True, override=True)

SPOTIPY_CLIENT_ID = os.environ.get('SPOTIPY_CLIENT_ID')
SPOTIPY_CLIENT_SECRET = os.environ.get('SPOTIPY_CLIENT_SECRET')
SPOTIPY_REDIRECT_URI = os.environ.get('SPOTIPY_REDIRECT_URI')
